+++ pyproject.toml
@@ -75,7 +75,7 @@ testing_minimal = [
   "hypothesis>=6.148.9",
   "z3-solver<4.15.4",  # 4.15.4 has a segfault when creating many z3.Context()
 ]
-testing_unit = ["tinygrad[testing_minimal]", "tqdm", "safetensors", "tabulate", "openai", "gguf>=0.18", "capstone"]
+testing_unit = ["tinygrad[testing_minimal]", "tqdm", "safetensors", "tabulate", "openai", "gguf>=0.18"]
 testing = [
   "tinygrad[testing_unit]",
   "pillow",
@@ -93,6 +93,7 @@ testing = [
   "networkx",
   "nibabel",
   "bottle",
+  "capstone",
   "pycocotools",
   "boto3",
   "pandas",
